Refactor cli into lib layers and add `vm run` witness generation
Split the cli crate into three layers so the binary and a future SDK can
share logic: `common` (helpers), `ops` (clap-free typed operations), and
`commands` (thin clap definitions).

Add `vm run` subcommand that generates the witness (Footprints/traces) by
executing the entry function in-process via the Move VM + OnDiskStateView,
deriving the entry from `[circuit.<name>].entry` in Move.toml. Enables the
`footprint` feature on move-vm-runtime. Execution is a dry run; freshly
compiled modules always overwrite storage so stale bytecode is never run.

Move per-operation flags (--params-path/--pubs-indices/--kzg) onto the
prove/verify/test subcommands so `run` no longer requires an SRS; keep
--package-path/--circuit-name shared. Update README accordingly.
