+++ cli/src/commands/poseidon.rs
@@ -1,10 +1,9 @@
+// Copyright (c) zkMove Authors
+
+use crate::ops;
 use anyhow::Result;
 use clap::Parser;
-use halo2_proofs::halo2curves::{bn256::Fr, ff::PrimeField};
 use log::info;
-use move_core_types::u256::U256;
-
-const DOMAIN_SPEC: u64 = 1; // Domain spec for Poseidon hash
 
 #[derive(Parser)]
 #[command(about = "Utility for poseidon hash")]
@@ -26,11 +25,8 @@ impl PoseidonCommand {
             let mut rng = rand::thread_rng();
             rng.gen()
         });
-        let hash_result = poseidon_base::Hashable::hash_with_domain(
-            [Fr::from_u128(self.value), Fr::from_u128(nonce)],
-            Fr::from(DOMAIN_SPEC),
-        );
-        let hash_val = U256::from_le_bytes(&hash_result.to_repr().as_ref().try_into()?);
+
+        let hash_val = ops::poseidon::poseidon_hash(self.value, nonce)?;
 
         info!("Value: {}", self.value);
         info!("Nonce: {}", nonce);
