+++ cli/src/ops/mod.rs
@@ -1,0 +1,12 @@
+// Copyright (c) zkMove Authors
+
+//! Reusable, clap-free operations shared by the CLI binary and (future) SDK callers.
+//!
+//! Each module takes typed inputs and returns typed outputs, leaving argument parsing
+//! and most file IO to the `commands` layer.
+
+pub mod poseidon;
+pub mod prove;
+pub mod run;
+pub mod test_verifier;
+pub mod verify;
